move public dir outside src, small fixes in various places

+++ src/models/message.ts
@@ -22,7 +22,7 @@ const Message = function (sequelize: Sequelize, DataTypes: DataTypes) {
         authorId: DataTypes.INTEGER,
         author: DataTypes.STRING,
         message: DataTypes.STRING,
-        roomId: DataTypes.INTEGER,
+        roomId: DataTypes.BIGINT,
         isGeneral: DataTypes.BOOLEAN
     });
